// src/types/dictionaries/res/dict.ts
import { Language } from '../i18n'
import ResponseEn from './en'
import ResponseKr from './kr'

export interface ResponseTextInfo {
	status: number // HTTP status code
	code: number // Business-level code
	message: string // Human-readable message
}

export type ResponseKey =
	| 'RESPONSE_USER_LIST_SUCCESS'
	| 'RESPONSE_USER_LIST_FAILED'
	| 'RESPONSE_USER_DETAIL_SUCCESS'
	| 'RESPONSE_USER_DETAIL_FAILED'
	| 'RESPONSE_USER_CREATE_SUCCESS'
	| 'RESPONSE_USER_CREATE_FAILED'
	| 'RESPONSE_USER_UPDATE_SUCCESS'
	| 'RESPONSE_USER_UPDATE_FAILED'
	| 'RESPONSE_USER_DELETE_SUCCESS'
	| 'RESPONSE_USER_DELETE_FAILED'
	| 'RESPONSE_LOGIN_SUCCESS'
	| 'RESPONSE_LOGIN_FAILED'
	| 'RESPONSE_LOGOUT_SUCCESS'
	| 'RESPONSE_LOGOUT_FAILED'
	| 'RESPONSE_JOIN_SUCCESS'
	| 'RESPONSE_JOIN_FAILED'
	| 'RESPONSE_SESSION_CHECK_SUCCESS'
	| 'RESPONSE_SESSION_CHECK_FAILED'
	| 'RESPONSE_PROFILE_CREATE_SUCCESS'
	| 'RESPONSE_PROFILE_CREATE_FAILED'
	| 'RESPONSE_PROFILE_CHANGE_SUCCESS'
	| 'RESPONSE_PROFILE_CHANGE_FAILED'
	| 'RESPONSE_PROFILE_CHECK_SUCCESS'
	| 'RESPONSE_PROFILE_CHECK_FAILED'
	| 'RESPONSE_PROFILE_DELETE_SUCCESS'
	| 'RESPONSE_PROFILE_DELETE_FAILED'
	| 'RESPONSE_TAG_DUPLICATION_CHECK_SUCCESS'
	| 'RESPONSE_TAG_DUPLICATION_CHECK_FAILED'
	| 'RESPONSE_AUTH_USER_PROFILE_SUCCESS'
	| 'RESPONSE_AUTH_USER_PROFILE_FAILED'
	| 'RESPONSE_FRIEND_LIST_SUCCESS'
	| 'RESPONSE_FRIEND_LIST_FAILED'
	| 'RESPONSE_FRIEND_ADD_SUCCESS'
	| 'RESPONSE_FRIEND_ADD_FAILED'
	| 'RESPONSE_FRIEND_FILTER_SUCCESS'
	| 'RESPONSE_FRIEND_FILTER_FAILED'
	| 'RESPONSE_FRIEND_REQUEST_SUCCESS'
	| 'RESPONSE_FRIEND_REQUEST_FAILED'
	| 'RESPONSE_MESSAGE_LIST_SUCCESS'
	| 'RESPONSE_MESSAGE_LIST_FAILED'
	| 'RESPONSE_MESSAGE_SEND_SUCCESS'
	| 'RESPONSE_MESSAGE_SEND_FAILED'
	| 'RESPONSE_MESSAGE_DETAIL_SUCCESS'
	| 'RESPONSE_MESSAGE_DETAIL_FAILED'
	| 'RESPONSE_ROOM_LIST_SUCCESS'
	| 'RESPONSE_ROOM_LIST_FAILED'
	| 'RESPONSE_ROOM_DETAIL_SUCCESS'
	| 'RESPONSE_ROOM_DETAIL_FAILED'
	| 'RESPONSE_ROOM_CREATE_SUCCESS'
	| 'RESPONSE_ROOM_CREATE_FAILED'
	| 'RESPONSE_SEARCH_SUCCESS'
	| 'RESPONSE_SEARCH_FAILED'
	| 'RESPONSE_SEARCH_USER_SUCCESS'
	| 'RESPONSE_SEARCH_USER_FAILED'
	| 'RESPONSE_BAD_REQUEST'
	| 'RESPONSE_UNAUTHORIZED'
	| 'RESPONSE_INTERNAL_SERVER_ERROR'

export type ResponseDictionaryType = Record<ResponseKey, ResponseTextInfo>

export const ResponseDictionary: Record<Language, ResponseDictionaryType> = {
	en: ResponseEn,
	kr: ResponseKr,
}